Minor fixes, added check if there was an error while uploading

+++ main.py
@@ -11,14 +11,14 @@ import instaloader
 
 
 # https://www.youtube.com/watch?v=YD_N6Ffoojw
-# https://sssinstagram.com/
 
 #------------------
 def remove_utc_files():
     for file in os.listdir("."):
         if "UTC" in file:
             try:
-                os.remove(file)
+                if ".mp4" not in file:
+                    os.remove(file)
             except Exception as e:
                 print(f"Could not remove {file}: {e}")
 #------------------
@@ -31,7 +31,7 @@ intents.members = True
 
 bot = commands.Bot(command_prefix='!shit', intents=intents)
 loader = instaloader.Instaloader(dirname_pattern=".")
-def test(str):
+def downloadFunction(str):
     shortcode = str.split("/")[-2]
     post = instaloader.Post.from_shortcode(loader.context, shortcode)
     loader.download_post(post, target=".")
@@ -53,13 +53,15 @@ async def on_message(message):
     if message.author == bot.user:
         return
     if "https://www.instagram.com/" in message.content.lower():
-        file2 = test(message.content.strip())
-        end = file2.split(".")[1]
-        await message.channel.send(file=discord.File(os.getcwd()+"/"+file2, filename="video."+end))
-        if os.path.exists(file2):
-            os.remove(file2)
+        file2 = downloadFunction(message.content.strip())
+        try:
+            end = file2.split(".")[1]
+            test = await message.reply(file=discord.File(os.getcwd()+"/"+file2, filename="video."+end))
             remove_utc_files()
-
+        except discord.HTTPException as e:
+            await message.reply("There was an error during the process. Maybe the file is too large?")
+        if os.path.exists(file2):
+                os.remove(file2)    
     await bot.process_commands(message)
